refactoring after merging with dev
