+++ src/main/java/az/gdg/msarticle/client/AuthenticationClient.java
@@ -2,8 +2,8 @@ package az.gdg.msarticle.client;
 
 import az.gdg.msarticle.model.client.auth.UserInfo;
 import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 
 import static az.gdg.msarticle.model.client.auth.HttpHeader.X_AUTH_TOKEN;
