+++ src/main/java/az/gdg/msarticle/client/TeamClient.java
@@ -1,9 +1,9 @@
 package az.gdg.msarticle.client;
 
-import java.util.List;
-
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
+
+import java.util.List;
 
 @FeignClient(value = "ms-team-client", url = "https://gdg-ms-team.herokuapp.com/api/members")
 public interface TeamClient {
