+++ src/main/java/az/gdg/msarticle/config/FeignConfig.java
@@ -2,7 +2,6 @@ package az.gdg.msarticle.config;
 
 import az.gdg.msarticle.client.TeamClient;
 import az.gdg.msarticle.service.MsAuthService;
-
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Configuration;
