+++ src/main/java/az/gdg/msarticle/controller/HealthController.java
@@ -1,7 +1,6 @@
 package az.gdg.msarticle.controller;
 
 import io.swagger.annotations.ApiOperation;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.CrossOrigin;
